requirements.txt
beautifulsoup4==4.9.3
fire==0.3.1
Jinja2==2.11.2
MarkupSafe==1.1.1
PyYAML==5.3.1
six==1.15.0
soupsieve==2.0.1
termcolor==1.1.0
